Changed threaded timer to schedule, option to enable/disable timelapse and add datetime to images, option to dotimelapse to specific folder

+++ TTimelapse.py
@@ -12,14 +12,20 @@ import subprocess
 sys.path.append('../pythonUtils')
 import TDateUtil as dateUtl
 from twython import Twython
+from PIL import Image # draw date and time on images
+from PIL import ImageFont
+from PIL import ImageDraw
 
 class Timelapse:   
 
     # TIMELAPSE constants
+    ENABLED = True
+    ADD_DATE_TIME = True
     TIME_TO_TIMELAPSE = "01:00:00"
     INTERVAL_TO_CAPTURE_IMG = 120   # time in seconds
     CAPTURE_IMG_RETRIES = 3
     FFMPEG_PROCESS_TIMEOUT = 600    # time in seconds
+    PENDING = ""
 
     # TWITTER constants
     TWITTER_ENABLED = False
@@ -53,10 +59,13 @@ class Timelapse:
             logger.warning("timelapse.ini not found, using default values")
         else:
             self.config.read("timelapse.ini")
+            Timelapse.ENABLED = self.config["TIMELAPSE"].getboolean("enabled")
+            Timelapse.ADD_DATE_TIME = self.config["TIMELAPSE"].getboolean("add_date_time")
             Timelapse.TIME_TO_TIMELAPSE = str(self.config["TIMELAPSE"]["time_to_timelapse"])
             Timelapse.INTERVAL_TO_CAPTURE_IMG = int(self.config["TIMELAPSE"]["interval_to_capture_img"])
             Timelapse.CAPTURE_IMG_RETRIES = int(self.config["TIMELAPSE"]["capture_img_retries"])
             Timelapse.FFMPEG_PROCESS_TIMEOUT = int(self.config["TIMELAPSE"]["ffmpeg_process_timeout"])
+            Timelapse.PENDING = str(self.config["TIMELAPSE"]["pending"])
             Timelapse.TWITTER_ENABLED = self.config["TWITTER"].getboolean("enabled")
             if(Timelapse.TWITTER_ENABLED == True):
                 Timelapse.TWITTER_CONSUMER_KEY = str(self.config["TWITTER"]["consumer_key"])
@@ -71,7 +80,8 @@ class Timelapse:
         self.url = url
         self.schedule=schedule
         # Set timer to capture the images
-        self.timer = threading.Timer(Timelapse.INTERVAL_TO_CAPTURE_IMG, self.captureImage) 
+        # Not doing by a threaded timer, but by schedule module
+        # self.timer = threading.Timer(Timelapse.INTERVAL_TO_CAPTURE_IMG, self.captureImage) 
         # Set hour for transform images into timelapse and upload
         self.schedule.every().day.at(Timelapse.TIME_TO_TIMELAPSE).do(self.doTimelapse)
         # to query ffmpegProcess on doTimelapse
@@ -94,13 +104,12 @@ class Timelapse:
         
         # The "output" folder is ok, now we need to start capturing images
         if(self.directoryInfo == Timelapse.IS_WRITABLE):         
-            # A frame every 10min, so we'll have 10fph, 240fpd -> 1h of the day = 1s = 10 imgs   
+            # A frame every 2min, so we'll have 30fph, 30*24 fpd -> 1h of the day = 1s = 30 imgs   
             self.logger.debug("Timer started, interval=" + str(Timelapse.INTERVAL_TO_CAPTURE_IMG) + "s")
-            self.timer.start() 
+            self.schedule.every(Timelapse.INTERVAL_TO_CAPTURE_IMG).seconds.do(self.captureImage).tag("captureImage")           
             
-    def __del__(self):
-        self.timer.cancel()
-        schedule.clear()
+    def __del__(self):      
+        self.schedule.clear()
     
     def timerCancel(self):
         self.timer.cancel()
@@ -153,26 +162,72 @@ class Timelapse:
                 except urllib.error.ContentTooShortError as e:
                     self.logger.error("ContentTooShortError=" + str(e))
                 self.logger.warning("Capture Img retry " + str(retry+1) + "/" + str(Timelapse.CAPTURE_IMG_RETRIES))
-        self.timer = threading.Timer(Timelapse.INTERVAL_TO_CAPTURE_IMG, self.captureImage)
-        self.timer.start()
+        # Not doing by threaded timer. Issues with deleting it, verify!! 
+        # Now doing by module schedule
+        #self.timer = threading.Timer(Timelapse.INTERVAL_TO_CAPTURE_IMG, self.captureImage)
+        #self.timer.start()
+    
+    # TODO, pending folders with errors to try timelapse/upload later
+    # def appendToPending(self, dateFolder):
+    #     pending = str(self.config["TIMELAPSE"]["pending"]).split(",")
+    #     pending = ",".join(pending)
+    #     self.config["TIMELAPSE"]["pending"] = pending
+    #     with open('timelapse.ini', 'w') as configfile:
+    #         self.config.write(configfile)
 
-    def doTimelapse(self):        
+    def enableTimelapse(self):
+        Timelapse.ENABLED = True
+        self.logger.debug("enableTimelapse enabled=" + str(Timelapse.ENABLED))
+
+    def disableTimelapse(self):
+        Timelapse.ENABLED = False
+        self.logger.debug("disableTimelapse enabled=" + str(Timelapse.ENABLED))
+
+    def statusTimelapse(self):        
+        self.logger.debug("statusTimelapse enabled=" + str(Timelapse.ENABLED))   
+        return Timelapse.ENABLED
+
+    def doTimelapse(self, folder2timelapse=None):
+        if(Timelapse.ENABLED == True and folder2timelapse==None):
+            self.logger.debug("Timelapse job is not enabled. To do a timelapse, give the folder name as argument in doTimelapse()")
+            return
+        
+        # No folder as argument was given, so try the yestareday folder
+        self.folder2timelapse = folder2timelapse
+        if(self.folder2timelapse == None):
+            self.folder2timelapse = dateUtl.getYesterdayTimeStamp()
+
+        # The folder exists? If not, return...
+        if(os.path.exists(self.directory + "/" + self.folder2timelapse) == False):
+            self.logger.debug("No folder with images from yesterday to timelapse")
+            return 
+
         # Remove if already have a temporary folder output/timelapse_*
         fileList = glob.glob(self.directory + "/timelapse_*")
         for filePath in fileList:
-            shutil.rmtree(filePath)
+            shutil.rmtree(filePath)                   
 
         # Copy the last folder to a temporary location: output/timelapse_{YESTERDAY_TIMESTAMP}
-        subprocess.call(["cp", "-r", "output/" + dateUtl.getYesterdayTimeStamp(), "output/timelapse_" + dateUtl.getYesterdayTimeStamp()])
+        subprocess.call(["cp", "-r", self.directory+ "/" + self.folder2timelapse, self.directory+"/timelapse_" + self.folder2timelapse])
+
+        # Added date and time to all images in folder
+        if(Timelapse.ADD_DATE_TIME == True):
+            self.logger.debug("Adding datetime to images on temporary folder=" + self.directory+"/timelapse_" + self.folder2timelapse)
+            self.addDateTimeToImages(self.directory+"/timelapse_" + self.folder2timelapse)
 
         # Rename all files to a sequence img_{0000-9999}.jpg
+        self.logger.debug("Renaming files in " + self.directory+"/timelapse_" + self.folder2timelapse)
+        dirList = os.listdir(self.directory+"/timelapse_" + self.folder2timelapse)
+        dirList.sort() # to ensure the order of files from YYYYMMDD000000 to YYYYMMDD235959
         i=0
-        for filename in os.listdir("output/timelapse_" + dateUtl.getYesterdayTimeStamp()):
-            src=self.directory+"/timelapse_" + dateUtl.getYesterdayTimeStamp() + "/" + filename
-            dst=self.directory+"/timelapse_{}/img_{:04d}.jpg".format(dateUtl.getYesterdayTimeStamp(), i)   
-            os.rename(src, dst)  
-            i+=1
-        self.logger.debug("Renamed %s files in %s", str(i), self.directory+"/timelapse_" + dateUtl.getYesterdayTimeStamp())
+        for filename in dirList:
+            if filename.endswith(".jpg"):
+                src=self.directory+"/timelapse_" + self.folder2timelapse + "/" + filename
+                dst=self.directory+"/timelapse_{}/img_{:04d}.jpg".format(self.folder2timelapse, i) 
+                # self.logger.debug("src=" + src + ", dest=" + dst)
+                os.rename(src, dst)  
+                i+=1
+        self.logger.debug("Renamed %s files in %s", str(i), self.directory+"/timelapse_" + self.folder2timelapse)
 
         # Do timelapse with ffmpeg. These configs are specials to fit the video requirement for twitter
         # ffmpeg -f image2 -framerate 30 -i img_%04d.jpg -vcodec libx264 -pix_fmt yuv420p -strict -2 -acodec aac -q:v 1 test.mp4
@@ -187,13 +242,13 @@ class Timelapse:
         #   -test.avi, output file
         self.ffmpegProcess = subprocess.Popen([ "ffmpeg", "-f", "image2", 
                                                 "-framerate", "30", 
-                                                "-i", self.directory+"/timelapse_"+dateUtl.getYesterdayTimeStamp()+"/img_%04d.jpg", 
+                                                "-i", self.directory+"/timelapse_"+self.folder2timelapse+"/img_%04d.jpg", 
                                                 "-vcodec", "libx264",
                                                 "-pix_fmt", "yuv420p",
                                                 "-strict", "-2",
                                                 "-acodec", "aac",
                                                 "-q:v", "1",
-                                                self.directory+"/timelapse_"+dateUtl.getYesterdayTimeStamp()+"/"+dateUtl.getYesterdayTimeStamp()+".mp4"]
+                                                self.directory+"/timelapse_"+self.folder2timelapse+"/"+self.folder2timelapse+".mp4"]
                                                 ,shell=False
                                                 ,stderr=subprocess.DEVNULL # for no stdout on console
                                                 ,stdout=subprocess.DEVNULL)
@@ -213,7 +268,7 @@ class Timelapse:
             self.logger.debug("ret ffmpegProcess=%s", str(poll))
             self.schedule.clear("ffmpegProcess")
             if(poll == 0):
-                videoPath=self.directory+"/timelapse_"+dateUtl.getYesterdayTimeStamp()+"/"+dateUtl.getYesterdayTimeStamp()+".mp4"
+                videoPath=self.directory+"/timelapse_"+self.folder2timelapse+"/"+self.folder2timelapse+".mp4"
                 if(Timelapse.TWITTER_ENABLED == True):                
                     self.logger.debug("ffmpeg process ended, uploading it to twitter, videoPath=%s", videoPath)                
                     self.uploadVideoTwitter(videoPath)
@@ -234,5 +289,28 @@ class Timelapse:
         
         video = open(videoPath, 'rb')
         response = twitter.upload_video(media=video, media_type='video/mp4')
-        twitter.update_status(status="A beauty timelapse from Araras, Sao Paulo, Brazil. It's from " + dateUtl.getYesterdayTimeStamp("%d/%m/%Y") + "."
+        twitter.update_status(status="A beauty timelapse from Araras, Sao Paulo, Brazil. It's from " + self.folder2timelapse + "."
                               ,media_ids=[response['media_id']])
+    
+    def addDateTimeToImages(self, folderImages):
+        # REF: https://www.amphioxus.org/content/timelapse-time-stamp-overlay
+        font = ImageFont.truetype("./HelveticaNeue.dfont", 43)
+        fontsmall = ImageFont.truetype("./HelveticaNeue.dfont", 32)
+        fontcolor = (238, 161, 6)
+        counter = 0
+        directory = folderImages
+        for file in os.listdir(directory):
+            if file.endswith(".jpg"):        
+                counter += 1	
+                #print("Image {0}: {1}".format(counter, file))
+
+                # Files are named like YYYYMMDDhhmmss.jpg, extract date and time
+                date, time = dateUtl.getDateTimeFromFilename(file)
+
+                # Get a drawing context        
+                img = Image.open(directory + "/" + file)
+                draw = ImageDraw.Draw(img)
+                draw.text((img.width-220, img.height-120), date, fontcolor, font=fontsmall)
+                draw.text((img.width-220, img.height-90), time, fontcolor, font=font)
+                img.save(directory + "/" + file)
+        self.logger.debug("Added datetime to " + str(counter) + " images")
